@@ -10,7 +10,7 @@ apply from: 'tools.gradle'
 
 
 android {
-    compileSdkVersion 28
+    compileSdkVersion 30
     signingConfigs {
         release {
             storeFile rootProject.file("app/jooxsdkdemo.keystore")
@@ -32,7 +32,7 @@ android {
     defaultConfig {
         applicationId "com.tencent.joox.sdk"
         minSdkVersion 16
-        targetSdkVersion 28
+        targetSdkVersion 30
         versionCode 1
         versionName "1.0"
         testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
@@ -91,12 +91,12 @@ repositories {
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation"org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
-    implementation 'com.android.support:appcompat-v7:28.0.0'
-    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
+    implementation 'androidx.appcompat:appcompat:1.2.0'
+    implementation 'androidx.constraintlayout:constraintlayout:2.0.0'
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'com.android.support.test:runner:1.0.2'
     androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
-    implementation 'com.android.support:recyclerview-v7:28.0.0'
+    implementation 'androidx.recyclerview:recyclerview:1.2.0'
     api(name:'JOOX_SDK_2.1.0.5',ext:'aar')
     api(name:'OpenSDK-release',ext:'aar')
     api(name: 'fingerprint_pure_lib-1.0.2-release', ext: 'aar')
