@@ -1,18 +1,17 @@
 package com.tencent.joox.sdk.songlisttest;
 
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
 import android.widget.TextView;
 import android.widget.Toast;
-
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.joox.sdklibrary.SDKInstance;
@@ -21,7 +20,6 @@ import com.tencent.joox.sdk.AppPlayManager;
 import com.tencent.joox.sdk.R;
 import com.tencent.joox.sdk.RequestParamBuilder;
 import com.tencent.joox.sdk.TrackItem;
-
 import java.util.List;
 
 public class ArtistsSongsActivity extends AppCompatActivity implements AppPlayManager.PlayerListener {
