@@ -2,11 +2,10 @@ package com.tencent.joox.sdk.songlisttest;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.widget.EditText;
-
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
 import com.tencent.joox.sdk.R;
 
 public class InputSongListIdActivity extends AppCompatActivity {
