@@ -1,11 +1,6 @@
 package com.tencent.joox.sdk.songlisttest;
 
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -13,7 +8,11 @@ import android.view.ViewGroup;
 import android.widget.ImageView;
 import android.widget.TextView;
 import android.widget.Toast;
-
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.joox.sdklibrary.SDKInstance;
@@ -22,7 +21,6 @@ import com.tencent.joox.sdk.AppPlayManager;
 import com.tencent.joox.sdk.R;
 import com.tencent.joox.sdk.RequestParamBuilder;
 import com.tencent.joox.sdk.TrackItem;
-
 import java.util.List;
 
 public class SongListActivity extends AppCompatActivity implements AppPlayManager.PlayerListener {
@@ -118,14 +116,14 @@ public class SongListActivity extends AppCompatActivity implements AppPlayManage
         }
 
         @Override
-        public void onBindViewHolder(@NonNull CategoryViewHolder categoryViewHolder, int i) {
-            TrackItem categoriesBean = categories.get(i);
+        public void onBindViewHolder(@NonNull CategoryViewHolder categoryViewHolder, int position) {
+            TrackItem categoriesBean = categories.get(position);
             categoryViewHolder.textView.setText(categoriesBean.getName() + ":" + categoriesBean.getTrack_label_flag());
             categoryViewHolder.itemView.setOnClickListener(new View.OnClickListener() {
                 @Override
                 public void onClick(View v) {
                     AppPlayManager.getInstance().setPlayList(categories);
-                    AppPlayManager.getInstance().playIndex(i);
+                    AppPlayManager.getInstance().playIndex(position);
                 }
             });
         }
